[Refactoring ] 프로필 이미지를 받아오는 로직 useEffect => hook으로 분리
